Rename settings editor -> hotkeys editor
